Added map component changes
