+++ components/src/components/Map/Common/Constants.d.ts
@@ -1,0 +1,4 @@
+export declare abstract class Constants {
+    static readonly DEFAULT_HELP_LINK: string;
+    static readonly ADV: string;
+}
