+++ components/src/components/Map/Common/Constants.js
@@ -1,0 +1,4 @@
+export class Constants {
+    static DEFAULT_HELP_LINK = 'https://github.com/bcgov/common-hosted-form-service/wiki';
+    static ADV = '';
+}
