--- components/src/components/Map/Common/Constants.ts
@@ -1,4 +1,0 @@
-export abstract class Constants {
-  static readonly DEFAULT_HELP_LINK: string = 'https://github.com/bcgov/common-hosted-form-service/wiki';
-  static readonly ADV: string = '';
-}
