+++ components/src/components/Map/Component.form.d.ts
@@ -1,0 +1,1 @@
+export default function (...extend: any[]): any;
