+++ components/src/components/Map/Component.form.js
@@ -1,13 +1,10 @@
 import baseEditForm from 'formiojs/components/_classes/component/Component.form';
-
 import EditData from './editForm/Component.edit.data';
 import EditDisplay from './editForm/Component.edit.display';
 import EditValidation from './editForm/Component.edit.validation';
-
 import SimpleApi from '../Common/Simple.edit.api';
 import SimpleConditional from '../Common/Simple.edit.conditional';
-
-export default function(...extend) {
+export default function (...extend) {
     return baseEditForm([
         EditDisplay,
         {
