+++ components/src/components/Map/editForm/Component.edit.data.d.ts
@@ -1,0 +1,29 @@
+declare const _default: ({
+    weight: number;
+    type: string;
+    label: string;
+    tooltip: string;
+    key: string;
+    input: boolean;
+    placeholder?: undefined;
+} | {
+    type: string;
+    label: string;
+    key: string;
+    weight: number;
+    placeholder: string;
+    tooltip: string;
+    input: boolean;
+} | {
+    weight: number;
+    type: string;
+    label: string;
+    key: string;
+    tooltip: string;
+    input: boolean;
+    values: {
+        value: string;
+        label: string;
+    }[];
+})[];
+export default _default;
