+++ components/src/components/Map/editForm/Component.edit.validation.d.ts
@@ -1,0 +1,26 @@
+declare const _default: ({
+    weight: number;
+    type: string;
+    defaultValue: boolean;
+    label: string;
+    tooltip: string;
+    key: string;
+    input: boolean;
+} | {
+    weight: number;
+    type: string;
+    label: string;
+    tooltip: string;
+    key: string;
+    input: boolean;
+    placeholder?: undefined;
+} | {
+    weight: number;
+    key: string;
+    label: string;
+    placeholder: string;
+    type: string;
+    tooltip: string;
+    input: boolean;
+})[];
+export default _default;
